fix(wrpc) do a pcall for all export_deflated_reconfigure_payload calls
We are already wrapping some calls to
`export_deflated_reconfigure_payload()` inside a pcall in the
`wrpc_control_plane.lua` file. This change is doing a pcall in all the
remaining calls to `export_deflated_reconfigure_payload()` in this file
to avoid the CP crash whenever we find errors during initialization of
modules for example.